Fix Deletion of User to work when he's not a Player

+++ src/main/java/ch/uzh/ifi/seal/soprafs20/service/LobbyService.java
@@ -429,19 +429,21 @@ public class LobbyService
     }
 
     public void removeFromLobbyAndDeletePlayer(User toDeleteUser){
-        Player player = playerService.getPlayerByToken(toDeleteUser.getToken());
-        Lobby lobbyToRemovePlayer = null;
-        List<Lobby> lobbies= this.getLobbies();
-        for(Lobby lobby:lobbies){
-            Set<Player> players = lobby.getPlayers();
-            if(players.contains(player)){
-                lobbyToRemovePlayer = lobby;
-                break;
+        if (playerService.doesPlayerWithTokenExist(toDeleteUser.getToken())) {
+            Player player = playerService.getPlayerByToken(toDeleteUser.getToken());
+            Lobby lobbyToRemovePlayer = null;
+            List<Lobby> lobbies= this.getLobbies();
+            for(Lobby lobby:lobbies){
+                Set<Player> players = lobby.getPlayers();
+                if(players.contains(player)){
+                    lobbyToRemovePlayer = lobby;
+                    break;
+                }
             }
-        }
-        if(!lobbyToRemovePlayer.equals(null)) {
-            this.removePlayerFromLobby(lobbyToRemovePlayer.getId(), player.getId());
-            playerService.deletePlayer(player);
+            if(!lobbyToRemovePlayer.equals(null)) {
+                this.removePlayerFromLobby(lobbyToRemovePlayer.getId(), player.getId());
+                playerService.deletePlayer(player);
+            }
         }
     }
 }
